product join review + pagenation productRepository 테스트 완 (all product)

+++ nurim-spring/src/main/java/human/nurim_spring/dto/ProductListResDto.java
@@ -13,5 +13,5 @@ public class ProductListResDto {
     private String img;
     private Long discountRate;
     private Long scopeCount;
-    private Long scopeAvg;
+    private Double scopeAvg;
 }
